refactor: flat section structure
