+++ apps/frontend/src/common/ui/SectionHeader.vue
